requirements.txt
bleach==2.1.2
bokeh==0.12.14
boto3==1.5.26
botocore==1.8.40
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.2
django-allauth==0.35.0
django-rest-framework==0.1.0
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
html5lib==1.0.1
idna==2.6
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
mysqlclient==1.3.12
numpy==1.14.0
oauthlib==2.0.6
packaging==16.8
psycopg2==2.7.4
pyparsing==2.2.0
python-dateutil==2.6.1
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.12
sendgrid==5.3.0
six==1.11.0
tornado==4.5.3
urllib3==1.22
webencodings==0.5.1
